+++ NEWS.md
@@ -1,3 +1,5 @@
+# eulerr (development version)
+
 # eulerr 6.1.1
 
 ## Minor changes